Fix cycles identified by ratchet
