+++ ts/packages/aiclient/src/apiSettingsFromConfig.ts
@@ -31,7 +31,8 @@ import {
     createAzureTokenProvider,
 } from "./auth.js";
 import type { AzureApiSettings } from "./azureSettings.js";
-import { ApiSettings, ModelType } from "./openai.js";
+import { ModelType } from "./apiTypes.js";
+import type { ApiSettings } from "./openai.js";
 import type { OpenAIApiSettings } from "./openaiSettings.js";
 
 const azureTokenProvider = createAzureTokenProvider(
