+++ ts/packages/aiclient/src/ollamaModels.ts
@@ -11,11 +11,11 @@ import {
 import { getEnvSetting } from "./common.js";
 import { ChatModelWithStreaming, CompletionSettings } from "./models.js";
 import {
-    CommonApiSettings,
     CompletionUsageStats,
     EnvVars,
     ModelType,
-} from "./openai.js";
+} from "./apiTypes.js";
+import type { CommonApiSettings } from "./openai.js";
 import {
     callApi,
     callJsonApi,
